Ontology Initialization Test
Created ontoTest.py, which allows to update the Jena Fuseki server with the contents of the ontoTestAddition.txt file, as well as query the server after the fact to get all of the RDF triples in the server, then parse out which ones were initialized through the prior step and create a GraphML file of those.
Also modified KOIOS to allow for importing of an existing GraphML file, and changed the graph structure within KOIOS to match the ontology a little better (at least with regards to Item/ItemRole elements).