Actions now map a bit better to the onto + onto changes
Changed KOIOS to map actions a little more in line with the ontology.  Also modified the ontology to have a test situation description/transition description, as well as changing the RDF->graphML processor to ignore "type" triples.

That last bit introduces a potential bug where if an item is only declared as being of some type (:Item6 rdf:type :Item), it gets completely glossed over.