Just an initial pass at using OwlReady2 to instantiate task/item/itemRole
Seems to work - outputs an OWL file with the initialized stuff.  Going to have to see how to use that to update the server.
One problem I'm facing still is that adding properties works like "item.hasItemName.append()" - so I can't just set it to "item.ITEM_NAME_EDGE.append()" and just update the constant.  Going to try a few things to get around that for the sake of maintainability.